requirements.txt
requests
colorama
pyunpack
patool
dnspython
ipaddress
python-masscan
python-nmap
pythonping
